to move middleware.js to middleware folder and rename, and optimize
